use CLI KT for args from inputs
